🏰 permit null values for fields that support null.
Note that the front-end still needs to provide these values if the Place has a value for these columns or they will show up as changed.

Note the difference between this method and the placeRouter#patch method, all fields are optional there. Here only fields that support "null" are optional.