🏰 Avoid crashing when update to place is only updating relations.
What? When the Assocations form is saved the only updates are to non-column attributes so the actual body of the update ends up being empty.

In other contexts, I would consider that these kinds of updates should be made to another endpoint, but that would make generating a PlaceEdit record more complex.